requirements.txt
cherrypy
git+https://github.com/mongodb/mongo-python-driver#egg=mongo-python-driver
mako
markdown
gitpython
